move geo_shape query builder tests to spatial module
